Create an InMemorySecurityStore
