+++ leshan-server-core/src/main/java/org/eclipse/leshan/server/impl/FileSecurityStore.java
@@ -21,15 +21,7 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.ObjectInputStream;
 import java.io.ObjectOutputStream;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.concurrent.locks.Lock;
-import java.util.concurrent.locks.ReadWriteLock;
-import java.util.concurrent.locks.ReentrantReadWriteLock;
 
-import org.eclipse.leshan.server.security.EditableSecurityStore;
 import org.eclipse.leshan.server.security.NonUniqueSecurityInfoException;
 import org.eclipse.leshan.server.security.SecurityInfo;
 import org.eclipse.leshan.server.security.SecurityStore;
@@ -44,20 +36,9 @@ import org.slf4j.LoggerFactory;
  * server is restarted.
  * </p>
  */
-public class FileSecurityStore implements EditableSecurityStore {
+public class FileSecurityStore extends InMemorySecurityStore {
 
     private static final Logger LOG = LoggerFactory.getLogger(FileSecurityStore.class);
-
-    // lock for the two maps
-    private final ReadWriteLock readWriteLock = new ReentrantReadWriteLock();
-    private final Lock readLock = readWriteLock.readLock();
-    private final Lock writeLock = readWriteLock.writeLock();
-
-    // by client end-point
-    private Map<String, SecurityInfo> securityByEp = new HashMap<>();
-
-    // by PSK identity
-    private Map<String, SecurityInfo> securityByIdentity = new HashMap<>();
 
     // the name of the file used to persist the store content
     private final String filename;
@@ -74,61 +55,9 @@ public class FileSecurityStore implements EditableSecurityStore {
         filename = file;
         loadFromFile();
     }
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public SecurityInfo getByEndpoint(String endpoint) {
-        readLock.lock();
-        try {
-            return securityByEp.get(endpoint);
-        } finally {
-            readLock.unlock();
-        }
-    }
 
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public SecurityInfo getByIdentity(String identity) {
-        readLock.lock();
-        try {
-            return securityByIdentity.get(identity);
-        } finally {
-            readLock.unlock();
-        }
-    }
-
-    @Override
-    public Collection<SecurityInfo> getAll() {
-        readLock.lock();
-        try {
-            return Collections.unmodifiableCollection(securityByEp.values());
-        } finally {
-            readLock.unlock();
-        }
-    }
-
-    private SecurityInfo addToStore(SecurityInfo info) throws NonUniqueSecurityInfoException {
-        writeLock.lock();
-        try {
-            String identity = info.getIdentity();
-            if (identity != null) {
-                SecurityInfo infoByIdentity = securityByIdentity.get(info.getIdentity());
-                if (infoByIdentity != null && !info.getEndpoint().equals(infoByIdentity.getEndpoint())) {
-                    throw new NonUniqueSecurityInfoException("PSK Identity " + info.getIdentity() + " is already used");
-                }
-
-                securityByIdentity.put(info.getIdentity(), info);
-            }
-
-            SecurityInfo previous = securityByEp.put(info.getEndpoint(), info);
-
-            return previous;
-        } finally {
-            writeLock.unlock();
-        }
+    protected SecurityInfo addToStore(SecurityInfo info) throws NonUniqueSecurityInfoException {
+        return super.add(info);
     }
 
     @Override
@@ -147,13 +76,8 @@ public class FileSecurityStore implements EditableSecurityStore {
     public SecurityInfo remove(String endpoint) {
         writeLock.lock();
         try {
-            SecurityInfo info = securityByEp.get(endpoint);
+            SecurityInfo info = super.remove(endpoint);
             if (info != null) {
-                if (info.getIdentity() != null) {
-                    securityByIdentity.remove(info.getIdentity());
-                }
-                securityByEp.remove(endpoint);
-
                 saveToFile();
             }
             return info;
